requirements.txt
tweety-ns
kafka-python
pyspark
numpy
pandas
scikit-learn
elasticsearch